rename module to reflect endpoint naming
